requirements.txt
Django~=5.0.3
djangorestframework~=3.14.0
djangorestframework-gis~=1.0
setuptools~=69.1.1